Renamed the project from FSM4J to Infinite
The previous name of the project, FSM4J, was pretty boring.  The new
name, Infinite, has a cool ring to it seeing how it is a hierarchical
FINITE state machine library.
