+++ LICENSE
@@ -11,8 +11,8 @@ modification, are permitted provided that the following conditions are met:
   this list of conditions and the following disclaimer in the documentation
   and/or other materials provided with the distribution.
 
-* Neither the name of FSM4J nor the names of its contributors may be used to
-  endorse or promote products derived from this software without specific
+* Neither the name of Infinite nor the names of its contributors may be used
+  to endorse or promote products derived from this software without specific
   prior written permission.
 
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
