+++ pom.xml
@@ -2,12 +2,12 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.bnorm.fsm4j.users.${user.name}</groupId>
-    <artifactId>fsm4j</artifactId>
+    <groupId>com.bnorm.infinite.users.${user.name}</groupId>
+    <artifactId>infinite</artifactId>
     <version>1.0.0-SNAPSHOT</version>
-    <name>fsm4j</name>
-    <description>Finite State Machine for Java</description>
-    <url>https://github.com/bnormsoftware/fsm4j</url>
+    <name>Infinite</name>
+    <description>Infinite - A hierarchical finite state machine for Java</description>
+    <url>https://github.com/bnorm-software/infinite</url>
 
     <organization>
         <name>BNORM Software</name>
@@ -15,14 +15,14 @@
     </organization>
 
     <issueManagement>
-        <url>https://github.com/bnormsoftware/fsm4j/issues</url>
+        <url>https://github.com/bnorm-software/infinite/issues</url>
         <system>GitHub Issues</system>
     </issueManagement>
 
     <scm>
-        <url>https://github.com/bnormsoftware/fsm4j</url>
-        <connection>scm:git:git://github.com/bnormsoftware/fsm4j.git</connection>
-        <developerConnection>scm:git:git@github.com:bnormsoftware/fsm4j.git</developerConnection>
+        <url>https://github.com/bnorm-software/infinite</url>
+        <connection>scm:git:git://github.com/bnorm-software/infinite.git</connection>
+        <developerConnection>scm:git:git@github.com:bnorm-software/infinite.git</developerConnection>
     </scm>
 
     <developers>
