+++ src/main/java/com/bnorm/infinite/InternalState.java
@@ -1,4 +1,4 @@
-package com.bnorm.fsm4j;
+package com.bnorm.infinite;
 
 import java.util.Optional;
 import java.util.Set;
