+++ src/main/java/com/bnorm/infinite/InternalStateFactory.java
@@ -1,4 +1,4 @@
-package com.bnorm.fsm4j;
+package com.bnorm.infinite;
 
 /**
  * A factory interface for internal states.
