+++ src/main/java/com/bnorm/infinite/StateMachineBase.java
@@ -1,4 +1,4 @@
-package com.bnorm.fsm4j;
+package com.bnorm.infinite;
 
 import java.util.Collections;
 import java.util.LinkedHashSet;
