+++ src/main/java/com/bnorm/infinite/StateMachineFactory.java
@@ -1,4 +1,4 @@
-package com.bnorm.fsm4j;
+package com.bnorm.infinite;
 
 import java.util.Map;
 import java.util.Set;
