+++ src/main/java/com/bnorm/infinite/Transition.java
@@ -1,4 +1,4 @@
-package com.bnorm.fsm4j;
+package com.bnorm.infinite;
 
 /**
  * Simple interface that represents a transition between states.
