+++ src/main/java/com/bnorm/infinite/TransitionFactory.java
@@ -1,4 +1,4 @@
-package com.bnorm.fsm4j;
+package com.bnorm.infinite;
 
 /**
  * @author Brian Norman
