+++ src/main/java/com/bnorm/infinite/TransitionGuard.java
@@ -1,4 +1,4 @@
-package com.bnorm.fsm4j;
+package com.bnorm.infinite;
 
 /**
  * Simple interface that represents a transition guard.
